DBSClustering Visualisation
To improve the code, one could also run the code multiple times with different datasets to see how the algorithm behaves with different data, and how the choice of parameters changes with different datasets. And also using other clustering algorithm and comparing the results with DBSCAN to see which algorithm performs better for the given dataset.